requirements.txt
pybit
anthropic
python-dotenv
fastapi
uvicorn
python-telegram-bot